--- src/main/java/ru/levelp/at/lesson0708/api/model/MetaData.java
@@ -1,19 +1,0 @@
-package ru.levelp.at.lesson0708.api.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@ToString
-@EqualsAndHashCode
-@Builder
-public class MetaData {
-
-    private PaginationData pagination;
-}
