--- src/main/java/ru/levelp/at/lesson0708/api/model/PeopleListResponse.java
@@ -1,21 +1,0 @@
-package ru.levelp.at.lesson0708.api.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-import java.util.List;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@ToString
-@EqualsAndHashCode
-@Builder
-public class PeopleListResponse {
-
-    private List<PersonData> data;
-    private MetaData meta;
-}
