--- src/main/java/ru/levelp/at/lesson0708/api/serialize/deserialize/Person.java
@@ -1,52 +1,0 @@
-package ru.levelp.at.lesson0708.api.serialize.deserialize;
-
-import java.io.Serializable;
-
-public class Person implements Serializable {
-
-    private String firstName;
-    private String lastName;
-    private Integer age;
-
-    public Person() {
-    }
-
-    public Person(String firstName, String lastName, Integer age) {
-        this.firstName = firstName;
-        this.lastName = lastName;
-        this.age = age;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public Integer getAge() {
-        return age;
-    }
-
-    public void setAge(Integer age) {
-        this.age = age;
-    }
-
-    @Override
-    public String toString() {
-        return "Person{"
-            + "firstName='" + firstName + '\''
-            + ", lastName='" + lastName + '\''
-            + ", age=" + age
-            + '}';
-    }
-}
