+++ src/main/java/ru/levelp/at/lesson0709/api/model/ViolationData.java
@@ -1,0 +1,21 @@
+package ru.levelp.at.lesson0709.api.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@ToString
+@EqualsAndHashCode
+@Builder
+public class ViolationData {
+
+    private String code;
+    private String field;
+    private String detail;
+}
