+++ src/test/resources/ru/levelp/at/lesson0709/api/database/database.properties
@@ -1,0 +1,3 @@
+url=jdbc:postgresql://localhost:5432/person_profile
+username=person_profile
+password=password
